Moved TestExperiment.py inside src directory
